# 26_PINEAPPLE-main/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/ping": {
      "get": {
        "operationId": "AppController_ping",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/admin-test": {
      "get": {
        "operationId": "AppController_adminTest",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns an array of users."
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findById",
        "summary": "Get a user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single user."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update a user (Requires Administrator)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "summary": "Delete a user (Requires Administrator)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/gamification/leaderboard": {
      "get": {
        "operationId": "GamificationController_getLeaderboard",
        "summary": "Get the global leaderboard sorted by XP",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns an array of gamification stats sorted by XP."
          }
        },
        "tags": [
          "Gamification"
        ]
      }
    },
    "/gamification/{userId}/ledger": {
      "get": {
        "operationId": "GamificationController_getLedgerByUserId",
        "summary": "Get the gamification ledger for a specific user",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of ledger entries."
          }
        },
        "tags": [
          "Gamification"
        ]
      }
    },
    "/projects": {
      "get": {
        "operationId": "ProjectsController_findAll",
        "summary": "Get all projects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns an array of projects."
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "Create a new project (Requires Project Owner or Administrator)",
        "parameters": [
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "description": "ID of the user creating the project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The project has been successfully created."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findById",
        "summary": "Get a project by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single project."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "summary": "Update a project (Requires Project Owner or Administrator)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The project has been successfully updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_delete",
        "summary": "Delete a project (Requires Project Owner or Administrator)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/owner/{ownerId}": {
      "get": {
        "operationId": "ProjectsController_findByOwnerId",
        "summary": "Get projects by Owner ID",
        "parameters": [
          {
            "name": "ownerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of projects belonging to the owner."
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/tasks/project/{projectId}": {
      "get": {
        "operationId": "TasksController_findAllByProjectId",
        "summary": "Get all tasks for a specific project",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of tasks."
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findById",
        "summary": "Get a task by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single task."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "summary": "Update a task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The task has been successfully updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_delete",
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The task has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "summary": "Create a new task",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The task has been successfully created."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    }
  },
  "info": {
    "title": "TeamForge API",
    "description": "The Student Project Collaboration Platform REST API",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "Collaborator",
              "Project Owner",
              "Mentor",
              "Administrator",
              "Super User"
            ],
            "example": "Collaborator"
          },
          "skills": {
            "example": [
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "linkedIn": {
            "type": "string",
            "example": "https://linkedin.com/in/johndoe"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Suspended",
              "Warned"
            ],
            "default": "Active"
          },
          "flags": {
            "type": "boolean",
            "example": false,
            "default": false
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "Collaborator",
              "Project Owner",
              "Mentor",
              "Administrator",
              "Super User"
            ],
            "example": "Collaborator"
          },
          "skills": {
            "example": [
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "linkedIn": {
            "type": "string",
            "example": "https://linkedin.com/in/johndoe"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Suspended",
              "Warned"
            ],
            "default": "Active"
          },
          "flags": {
            "type": "boolean",
            "example": false,
            "default": false
          }
        }
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "AI Study Planner"
          },
          "description": {
            "type": "string",
            "example": "An intelligent study scheduling app."
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ],
            "example": "Medium"
          },
          "requiredSkills": {
            "example": [
              "React",
              "Python",
              "ML"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "duration": {
            "type": "string",
            "example": "3 Months"
          }
        },
        "required": [
          "title",
          "description",
          "difficulty",
          "requiredSkills",
          "duration"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "AI Study Planner"
          },
          "description": {
            "type": "string",
            "example": "An intelligent study scheduling app."
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ],
            "example": "Medium"
          },
          "requiredSkills": {
            "example": [
              "React",
              "Python",
              "ML"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "duration": {
            "type": "string",
            "example": "3 Months"
          }
        }
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "example": "proj-1"
          },
          "title": {
            "type": "string",
            "example": "Design Database Schema"
          },
          "description": {
            "type": "string",
            "example": "Create the ERD and define the tables for the backend."
          },
          "xpReward": {
            "type": "number",
            "example": 50,
            "description": "XP awarded upon completion"
          },
          "status": {
            "type": "string",
            "enum": [
              "To Do",
              "In Progress",
              "In Review",
              "Completed"
            ],
            "default": "To Do"
          },
          "assigneeId": {
            "type": "string",
            "example": "2"
          }
        },
        "required": [
          "projectId",
          "title",
          "description",
          "xpReward"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "example": "proj-1"
          },
          "title": {
            "type": "string",
            "example": "Design Database Schema"
          },
          "description": {
            "type": "string",
            "example": "Create the ERD and define the tables for the backend."
          },
          "xpReward": {
            "type": "number",
            "example": 50,
            "description": "XP awarded upon completion"
          },
          "status": {
            "type": "string",
            "enum": [
              "To Do",
              "In Progress",
              "In Review",
              "Completed"
            ],
            "default": "To Do"
          },
          "assigneeId": {
            "type": "string",
            "example": "2"
          }
        }
      }
    }
  }
}